[#55] Separate Data Detail Component
